Fix CombiningOutputComponentTest warning on signed vs unsigned compare.
Also, doxygen is warning about the parameter "combiner" for
CombiningOutputComponent::setCombiner.  I don't know why it's having a
problem, so I documented the parameter without using the param syntax.

+++ SurgSim/Input/UnitTests/CombiningOutputComponentTest.cpp
@@ -257,7 +257,7 @@ TEST(CombiningOutputComponentTest, Serialization)
 	auto outputs = element->getComponents<SurgSim::Input::OutputComponent>();
 	ASSERT_EQ(outputs.size(), 4);
 	std::shared_ptr<SurgSim::Input::CombiningOutputComponent> combiningOutputComponent;
-	for (int i = 0; i < outputs.size(); ++i)
+	for (size_t i = 0; i < outputs.size(); ++i)
 	{
 		combiningOutputComponent = std::dynamic_pointer_cast<SurgSim::Input::CombiningOutputComponent>(outputs[i]);
 		if (combiningOutputComponent != nullptr)
